Saving PvP battle and replay

+++ src/main/java/swcnoops/server/model/Earned.java
@@ -1,7 +1,7 @@
 package swcnoops.server.model;
 
 public class Earned {
-    public long contraband;
-    public long credits;
-    public long materials;
+    public int contraband;
+    public int credits;
+    public int materials;
 }
